requirements.txt
# FastAPI and web
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis and Celery
redis==5.0.1
celery[redis]==5.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client
httpx==0.27.0
aiohttp==3.9.3

# Web scraping
playwright==1.41.2
beautifulsoup4==4.12.3
lxml==5.1.0

# Excel handling
openpyxl==3.1.2
pandas==2.2.0
xlsxwriter==3.1.9

# Word documents
python-docx==1.1.0

# Telegram
python-telegram-bot==21.0.1

# File storage (S3/MinIO)
boto3==1.34.34
minio==7.2.3

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# Calendar integration
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
transliterate==1.10.2
python-slugify==8.0.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.4

# Logging
structlog==24.1.0
